Refactor muxers to adapt other muxers in enterprise
